added saving and retrieving zones to and from the db

+++ src/main/java/com/dempseywood/entity/repository/EquipmentRepository.java
@@ -3,5 +3,5 @@ package com.dempseywood.entity.repository;
 import com.dempseywood.entity.Equipment;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EquipmentRepository extends CrudRepository<Equipment, Long> {
+public interface EquipmentRepository extends CrudRepository<Equipment, Integer> {
 }
