+++ src/main/java/com/dempseywood/entity/repository/TrackerRepository.java
@@ -3,6 +3,6 @@ package com.dempseywood.entity.repository;
 import com.dempseywood.entity.Tracker;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TrackerRepository extends CrudRepository<Tracker, Long> {
+public interface TrackerRepository extends CrudRepository<Tracker, Integer> {
 
 }
